requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
livekit==0.11.0
livekit-api==1.1.0
livekit-agents==1.0.0
livekit-agents[google]==1.0.0
livekit-plugins-google-ai-dev-kit==1.0.0
python-dotenv==1.0.1